Mongo_You7v/requirements.txt
fastapi
uvicorn
mongoengine
passlib
bcrypt
python-multipart
python-jose
cryptography
